Rename the folder for IconBackground.
